+++ web/js/036-functions.js
@@ -427,13 +427,23 @@
 		}
 
 		if (isMapOrPopup) {
-			$("ul#right-menu li.hide-descriptions-and-tags").addClass("hidden");
+			$("ul#right-menu li.hide-descriptions").addClass("hidden");
 		} else {
-			$("ul#right-menu li.hide-descriptions-and-tags").removeClass("hidden");
-			if (env.options.hide_descriptions_and_tags)
-				$("ul#right-menu li.hide-descriptions-and-tags").addClass("selected");
+			$("ul#right-menu li.hide-descriptions").removeClass("hidden");
+			if (env.options.hide_descriptions)
+				$("ul#right-menu li.hide-descriptions").addClass("selected");
 			else
-				$("ul#right-menu li.hide-descriptions-and-tags").removeClass("selected");
+				$("ul#right-menu li.hide-descriptions").removeClass("selected");
+		}
+
+		if (isMapOrPopup) {
+			$("ul#right-menu li.hide-tags").addClass("hidden");
+		} else {
+			$("ul#right-menu li.hide-tags").removeClass("hidden");
+			if (env.options.hide_tags)
+				$("ul#right-menu li.hide-tags").addClass("selected");
+			else
+				$("ul#right-menu li.hide-tags").removeClass("selected");
 		}
 
 		if (
@@ -553,7 +563,8 @@
 
 		if (
 			$("ul#right-menu li.hide-title").hasClass("hidden") &&
-			$("ul#right-menu li.hide-descriptions-and-tags").hasClass("hidden") &&
+			$("ul#right-menu li.hide-descriptions").hasClass("hidden") &&
+			$("ul#right-menu li.hide-tags").hasClass("hidden") &&
 			$("ul#right-menu li.media-count").hasClass("hidden") &&
 			$("ul#right-menu li.spaced").hasClass("hidden") &&
 			$("ul#right-menu li.square-album-thumbnails").hasClass("hidden") &&
@@ -1250,10 +1261,21 @@
 		else
 			$(".title").removeClass("hidden-by-option");
 
-		if (env.options.hide_descriptions_and_tags)
+		if (env.options.hide_descriptions && env.options.hide_tags) {
 			$("#description").addClass("hidden-by-option");
-		else
+		} else {
 			$("#description").removeClass("hidden-by-option");
+
+			if (env.options.hide_descriptions)
+				$("#description-title, #description-text").addClass("hidden-by-option");
+			else
+				$("#description-title, #description-text").removeClass("hidden-by-option");
+
+			if (env.options.hide_tags)
+				$("#description-tags").addClass("hidden-by-option");
+			else
+				$("#description-tags").removeClass("hidden-by-option");
+		}
 
 		if (env.options.hide_bottom_thumbnails && (env.currentMedia != null || env.currentAlbum.isAlbumWithOneMedia())) {
 			$("#album-view").addClass("hidden-by-option");
@@ -1469,9 +1491,13 @@
 							if (titleCookie !== null)
 								env.options.hide_title = titleCookie;
 
-							var descriptionsCookie = Functions.getBooleanCookie("hideDescriptionsAndTags");
+							var descriptionsCookie = Functions.getBooleanCookie("hideDescriptions");
 							if (descriptionsCookie !== null)
-								env.options.hide_descriptions_and_tags = descriptionsCookie;
+								env.options.hide_descriptions = descriptionsCookie;
+
+							var tagsCookie = Functions.getBooleanCookie("hideTags");
+							if (tagsCookie !== null)
+								env.options.hide_tags = tagsCookie;
 
 							var bottomThumbnailsCookie = Functions.getBooleanCookie("hideBottomThumbnails");
 							if (bottomThumbnailsCookie !== null)